requirements.txt
fastapi
uvicorn
pypdf2
python-docx
pandas
openpyxl
ollama
langchain
chromadb
langchain-openai
langchain-community
langchain-chroma
streamlit
requests